requirements.txt
pdfplumber
pypdf
openai
streamlit
opencv-python
python-dotenv
pillow
